Fix reading ingredient tables from older XML dumps
It turns out, that my assumption that the XML attributes
are always in the same order was not correct, it was changed
somewhere along the way. The current code uses the attribute
names returned by xml_attr to fix this. Thus, ingredient
information should now also be correct for the first years.