Use "g_per_L" for rate_g_per_ha calculation when only dosage is given and "g_per_L" exists.
In general products containing an active substance contents in "g_per_L" are liquids and products containing an active substance contents only in "percent" but not in "g_per_L" are solids.

For the calculation of "rate_g_per_ha" previously the active ingredient content in "percent" was used for entries with only a dosage (no rate). This has now been changed so that "g_per_L" is used if present and "percent" is used if there is no "g_per_L" entry.