Correct the coding of varying effects in obligations
Using `PEST_PARTIAL_EFFECT` in the obligations table for cases where
varying effects are specified in obligations could be misleading, as
this is the code officially used for cases where a full effect could not
be demonstrated for a certain target organism (pest). Therefore,
The the column `partial_effect` was renamed to `varying_effect` and
it contains TRUE in cases where the obligation text indicates a varying
effect (obligation code 1770 and similar wordings used before).
Include some example code demonstrating the varying effect
specifications in obligations. Update help pages generated with roxygen.